requirements.txt
python-dotenv
google-generativeai
langchain
langchain-openai
PyPDF2
faiss-cpu
streamlit
-e . 